BUGFIX : could not run tuttle again after it had been killed
killing tuttle can happen : stopping a Jenkins job, kill -9 because it's not responding, etc.

It would cause running processes to stay without an end attribute (which would indicate that it is still running). Hence, when running tuttle again, it would not invalidate outputs, and outputs signatures would be DISCOVERED (it would rase an asset on invalidation.py line 188)